requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
tenacity>=8.2.0
pandas>=2.1.0
numpy>=1.26.0
sqlalchemy>=2.0.0
scikit-learn>=1.4.0
xgboost>=2.0.0
joblib>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
jinja2>=3.1.0
pytz>=2023.3
pytest>=7.4.0
pytest-mock>=3.12.0
responses>=0.24.0
pytest-cov>=4.1.0
